feat: extract dev server logic as Node.js API
