+++ lib/api.js
@@ -1,5 +1,6 @@
 const { PassThrough } = require('stream')
 const { build } = require('./build')
+const { dev: startDevServer } = require('./dev')
 
 function isDev () {
   return process.env.NODE_ENV !== 'production'
@@ -20,3 +21,4 @@ function prod (stream) {
 exports.dev = dev
 exports.prod = prod
 exports.build = build
+exports.startDevServer = startDevServer
