+++ lib/cli/build.js
@@ -31,7 +31,5 @@ exports.handler = (argv, debug = {}) => {
   assert(configPath, 'Config file is not found')
 
   const options = util.merge(argv, { config: configPath })
-  return build(options, {
-    console: debug.console
-  })
+  return build(options, debug)
 }
