+++ lib/cli/dev.js
@@ -1,14 +1,9 @@
 'use strict'
 
 const assert = require('assert')
-const url = require('url')
-const externalIp = require('../util').getIp()
-const loadConfig = require('../config').loadConfig
 const findConfig = require('../config').findConfig
-const DepResolver = require('../dep-resolver')
-const create = require('../externals/browser-sync')
-const createWatcher = require('../externals/watcher')
-const DevLogger = require('../loggers/dev-logger')
+const { dev } = require('../dev')
+const util = require('../util')
 
 exports.builder = {
   config: {
@@ -25,59 +20,11 @@ exports.builder = {
   }
 }
 
-exports.handler = (argv, debug) => {
-  debug = debug || {}
-
+exports.handler = (argv, debug = {}) => {
   const configPath = argv.config || findConfig(process.cwd())
   assert(configPath, 'Config file is not found')
   assert(!Number.isNaN(argv.port), '--port should be a number')
 
-  const config = loadConfig(configPath).extend({
-    port: argv.port,
-    basePath: argv['base-path']
-  })
-
-  // Logger
-  const logger = new DevLogger(config, {
-    console: debug.console
-  })
-
-  const resolver = DepResolver.create(config)
-
-  const bs = create(config, {
-    logLevel: 'silent',
-    middleware: logMiddleware,
-    open: !argv._debug // Internal
-  }, resolver)
-
-  bs.emitter.on('init', () => {
-    logger.startDevServer(argv.port, externalIp)
-  })
-
-  const watcher = createWatcher(config, resolver, (name, files, origin) => {
-    if (name === 'add') {
-      logger.addFile(origin)
-    } else if (name === 'change') {
-      logger.updateFile(origin)
-    }
-    bs.reload(files.map(resolveOutput))
-  })
-
-  return { bs, watcher }
-
-  function resolveOutput (inputName) {
-    const rule = config.findRuleByInput(inputName)
-
-    if (!rule) {
-      return inputName
-    } else {
-      return rule.getOutputPath(inputName)
-    }
-  }
-
-  function logMiddleware (req, res, next) {
-    const parsedPath = url.parse(req.url).pathname
-    logger.getFile(parsedPath)
-    next()
-  }
+  const options = util.merge(argv, { config: configPath })
+  return dev(options, debug)
 }
